Centralize accessing nb metadata
